Sanitize failed mvn build when releasing
Resolves #598

+++ release.sh
@@ -61,6 +61,12 @@ then
 fi
 
 mvn --batch-mode release:prepare release:perform -DreleaseVersion=${version} -Darguments="-Dgpg.passphrase=${gpg_pass}" 
+mvnRetVal=$?
+if [ $mvnRetVal -ne 0 ]; then
+    echo "Mvn build failed - please fix it and try to release again"
+    exit $mvnRetVal
+fi
+
 sed -i.bak "s/\(<version>\)[0-9]*\.[0-9]*\.[0-9]*\(<\/version>\)/\1${version}\2/" README.md && rm -f README.md.bak
 git commit -a -m "bump version"
 git push origin --tags HEAD
